pyproject.toml
[tool.poetry]
name = "fast-authentication"
version = "0.0.1"
description = "Extensible authentication library for FastAPI"
authors = ["vipierozan99 <victorpierozan@gmail.com>"]
readme = "README.md"
packages = [{ include = "fast_authentication" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.98.0"
python-jose = "^3.3.0"
rich = "^13.4.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"
uvicorn = "^0.22.0"
better-exceptions = "^0.3.3"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.isort]
profile = "black"
